requirements.txt
whisperx
python-magic
python-magic-bin
